Add support for manylinux1
Pypi forces to upload packages that are compatible with all linux, we
have modified the release workflow following the instruction for
compiling a wheel that would need to be compatible for any linux.

Because of that the release workflow has changed a bit, and Mac and
Linux follow two different paths for compiling the wheel.

Once the wheel is generated we test it by installing it into a new
environment.