Add support for trace healthiness of the nodes (#27)
When a node can not have a established TCP connection, at least one,
is marked as unhealthy, and when the TCP connection is stablished
again the node is marked again as healthy. The user has the capacity of
removing from the pool of elegible nodes for sending traffic those nodes
that have been marked as unhealthy.

Also adds support for cluster events for listening for the significant
events that happen at cluster level, for now only support for knowing
when a host has changed its healthy status.