Add new `min_connections` parameters
A new kewyword argument has been added for the the client factory for
telling the Emcache client that a minimum number of TCP connections must
be keep opened. By default is configured to 1 which provides backward
compatibility

This number should be kept in ideal circumstances, but when there are outages
or the memcache serverers can not be reached these number can not be guaranteed.